requirements.txt
python3-yaml 
python3-selenium